refactor: split account identity components
