+++ frontend/src/accounts/detail/AccountDetailPage.tsx
@@ -5,7 +5,7 @@ import { useAccount, type Account } from '@/api/accounts'
 import { useTaxAdvantagedCategory } from '@/api/taxAdvantagedCategories'
 import type { Transaction } from '@/api/transactions'
 import CreateTransactionModal from '@/components/CreateTransactionModal'
-import AccountIdentityCard from '@/accounts/detail/components/AccountIdentityCard'
+import AccountIdentityCard from '@/accounts/detail/components/identity/AccountIdentityCard'
 import BackLink from '@/accounts/detail/components/BackLink'
 import BalanceChartCard from '@/accounts/detail/components/balance-chart/BalanceChartCard'
 import EditAccountIdentityModal from '@/accounts/detail/components/EditAccountIdentityModal'
